pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "squire"
version = "0.1.0"
description = "An AI agent that monitors and manages your homelab."
authors = [{name = "Will Ahern"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.23.0",
    "litellm>=1.81.1",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "textual>=0.80.0",
    "rich>=13.0.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
]

[project.scripts]
squire = "squire.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.14.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/squire"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
